Removed unneeded cast functions
